requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
fpdf2>=2.7.0
streamlit-code-editor>=0.1.20
